add test to check add and remove functionally
