+++ src/style.scss
@@ -115,7 +115,7 @@ main {
         }
       }
 
-      .actions {
+      .delete-btn {
         border: none;
         background-color: transparent;
       }
